% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{basic_stats_rate}
\alias{basic_stats_rate}
\title{This function produces the number of assists and turnover rates per game for a chosen player}
\usage{
basic_stats_rate(player = "Kyrie Irving", data = DATASET)
}
\arguments{
\item{player}{Name of NBA Player with the default being Kyrie Irving}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
Text giving the number of assists and turnover rates for a chosen player
}
\description{
This function produces the number of assists and turnover rates per game for a chosen player
}
\examples{
basic_stats_rate(player = "Kyrie Irving",data=DATASET)
}
