% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleandata}
\alias{cleandata}
\title{This function cleans any dataset retrieved from \href{https://www.basketball-reference.com/leagues/NBA_2023_per_game.html#per_game_stats}{NBA website}.
by removing rows with NAs and recoding the Position variable}
\usage{
cleandata(data = DATASET)
}
\arguments{
\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
returns a cleaned up version of data
}
\description{
This function cleans any dataset retrieved from \href{https://www.basketball-reference.com/leagues/NBA_2023_per_game.html#per_game_stats}{NBA website}.
by removing rows with NAs and recoding the Position variable
}
\examples{
cleandata(data = DATASET )
}
