% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{free_throw_rate}
\alias{free_throw_rate}
\title{This function produces the rate at which a chosen player gets fouled to shoot a free throw per game}
\usage{
free_throw_rate(player = "Kevin Durant", data = DATASET)
}
\arguments{
\item{player}{Name of NBA Player with the default being Kevin Durant}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
Text giving the free throw rate per game for the chosen player
}
\description{
This function produces the rate at which a chosen player gets fouled to shoot a free throw per game
}
\examples{
free_throw_rate(player = "Kevin Durant",data=DATASET)
}
