% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_panels_plot}
\alias{get_panels_plot}
\title{This function produces a correlation panel plot for the NBA season data}
\usage{
get_panels_plot(data = DATASET)
}
\arguments{
\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
A correlation plot of the NBA season data
}
\description{
This function produces a correlation panel plot for the NBA season data
}
\examples{
get_panels_plot(data=DATASET)
}
