% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modelling}
\alias{modelling}
\title{This function fits a regression model to either the free throw percentage or the minutes played based on a particular
characteristic measure}
\usage{
modelling(response = "FT.", data = DATASET)
}
\arguments{
\item{response}{A response variable chosen as either "FT." representing Free Throw Percentage or "MP"
representing Minuted Played}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
Regression summary modelling the chosen response
}
\description{
This function fits a regression model to either the free throw percentage or the minutes played based on a particular
characteristic measure
}
\examples{
modelling(response = "FT.",data=DATASET)
}
