% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{player_basic_stats}
\alias{player_basic_stats}
\title{This function produces a table for a given player's basic statistics (position, team, field goals, total rebounds,
assists, steals, blocks, turnovers, and personal fouls).}
\usage{
player_basic_stats(player = "Stephen Curry", data = DATASET)
}
\arguments{
\item{player}{Name of NBA Player with the default being Stephen Curry}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
Table that summarizes basic statistics for chosen player.
}
\description{
This function produces a table for a given player's basic statistics (position, team, field goals, total rebounds,
assists, steals, blocks, turnovers, and personal fouls).
}
\examples{
player_basic_stats(player = "Stephen Curry", data=DATASET)
}
