% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_basic_stats}
\alias{plot_basic_stats}
\title{This function produces a plot to compare the chosen statistical measure for different positions}
\usage{
plot_basic_stats(measure = "PTS", data = DATASET)
}
\arguments{
\item{measure}{Characteristic to compare players based on dataset columns}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
A barplot comparing the chosen statistical measure across the five positions
}
\description{
This function produces a plot to compare the chosen statistical measure for different positions
}
\examples{
plot_basic_stats(measure = "PTS",data=DATASET)
}
