% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{position_best}
\alias{position_best}
\title{This function produces the names of the top specified numbers of players in a certain position
based on a particular chosen characteristic}
\usage{
position_best(position = "Center", measure = "PTS", topn = 3, data = DATASET)
}
\arguments{
\item{position}{Name of position ("Center","Point Guard","Shooting Guard","Power Forward","Small Forward")}

\item{measure}{Characteristic to compare players based on dataset columns}

\item{topn}{Integer representing the desired number of top players}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
topn x 4 table representing the top n players, positions, teams, and desired measurement
}
\description{
This function produces the names of the top specified numbers of players in a certain position
based on a particular chosen characteristic
}
\examples{
position_best(position = "Center",measure = "PTS",topn = 3,data=DATASET)
}
