% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{true_shooting_percentage}
\alias{true_shooting_percentage}
\title{This function produces the true shooting percentage for a player.}
\usage{
true_shooting_percentage(player = "LeBron James", data = DATASET)
}
\arguments{
\item{player}{Name of NBA Player with the default being LeBron James}

\item{data}{Dataframe of NBA statistics With the default being DATASET the 2022-2023 season NBA dataset}
}
\value{
Text giving the true shooting percentage for the chosen player
}
\description{
This function produces the true shooting percentage for a player.
}
\examples{
true_shooting_percentage(player = "LeBron James",data=DATASET)
}
